Fix bug, move code to migrate.py
